requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
decorator==5.1.1
Django==4.2.5
django-enumchoicefield==3.0.1
djangorestframework==3.14.0
executing==2.0.0
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg==3.1.12
psycopg2==2.9.8
psycopg2-binary==2.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.10.1
typing_extensions==4.8.0
wcwidth==0.2.8
